% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a3_hil}
\alias{a3_hil}
\title{Tideman a3_hil}
\format{
A data frame with attribute "nseats" = 7, consisting of 989
observations and 15 candidates.
}
\usage{
data(a3_hil)
}
\description{
This data is one of 87 sets of ballots from the \href{https://rangevoting.org/TidemanData.html}{Tideman data collection}, as curated by \href{https://rangevoting.org/}{The Center for Range Voting}.

This set of ballots was collected in 1987 by Nicolaus Tideman, with support
from NSF grant SES86-18328. "The data are records of ballots from elections
of British organizations (mostly trade unions using PR-STV or IRV voting)
in which the voters ranked the candidates. The data were gathered under a
stipulation that the organizations involved would remain anonymous."

The ballots were encoded in David Hill's format, and have been converted to
the preference-vector format of this package.  The file
\href{https://rangevoting.org/TiData/A3.HIL}{A4.HIL} contains eight blank ballot
papers (1, 616, 619, 620, 685, 686, 687, 688) which we have retained. This
set may be counted by \code{stv(a3_hil,nseats=attr(a3_hil,"nseats"))}.
}
\keyword{datasets}
