% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{approval}
\alias{approval}
\title{Count votes using the approval method}
\usage{
approval(votes, nseats = 1, fsep = "\\t", quiet = FALSE, ...)
}
\arguments{
\item{votes, nseats, fsep, quiet, ...}{undocumented}
}
\value{
undocumented
}
\description{
See https://arxiv.org/abs/2102.05801
}
