% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{as.SafeRankExpt}
\alias{as.SafeRankExpt}
\title{as.SafeRankExpt()}
\usage{
as.SafeRankExpt(df)
}
\arguments{
\item{df}{data.frame object with all required attributes of a SafeRankExpt}
}
\value{
SafeRankExpt object
}
\description{
as.SafeRankExpt()
}
