% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{assemble.args.for.check.score}
\alias{assemble.args.for.check.score}
\title{undocumented internal method}
\usage{
assemble.args.for.check.score(x, max.score = NULL, ...)
}
\arguments{
\item{x, max.score, ...}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
