% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{assemble.args.for.check.stv}
\alias{assemble.args.for.check.stv}
\title{undocumented internal method}
\usage{
assemble.args.for.check.stv(x, equal.ranking = FALSE, ...)
}
\arguments{
\item{x, equal.ranking, ...}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
