% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{backwards.tiebreak}
\alias{backwards.tiebreak}
\title{Undocumented internal method}
\usage{
backwards.tiebreak(prefs, icans, elim = TRUE)
}
\arguments{
\item{prefs}{undocumented}

\item{icans}{undocumented}

\item{elim}{undocumented}
}
\description{
Undocumented internal method
}
