% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{check.nseats}
\alias{check.nseats}
\title{parameter-checking method for nseats (internal)}
\usage{
check.nseats(
  nseats = NULL,
  ncandidates,
  default = 1,
  mcan = NULL,
  complete.ranking = FALSE
)
}
\arguments{
\item{nseats}{initially-specified number of seats to be filled in an election}

\item{ncandidates}{the number of candidates standing for election}

\item{default}{the return value of this function when nseats=NULL}

\item{mcan}{a deprecated name for nseats}

\item{complete.ranking}{when TRUE, the return value is in 1..ncandidates
When FALSE, the return value is in 1..ncandidates-1 (for backwards
compatibility)}
}
\value{
a valid non-NULL value for the number of seats to be filled
}
\description{
parameter-checking method for nseats (internal)
}
