% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRankings.R
\name{check.ranking}
\alias{check.ranking}
\title{check the validity of a partial ranking}
\usage{
check.ranking(r)
}
\arguments{
\item{r}{a numeric vector}
}
\value{
a partial ranking of the elements of `r`, using `ties.method="min"`
}
\description{
check the validity of a partial ranking
}
