% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{check.votes}
\alias{check.votes}
\title{undocumented internal method}
\usage{
check.votes(x, ..., quiet = FALSE)
}
\arguments{
\item{x, quiet, ...}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
