% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{check.votes.score}
\alias{check.votes.score}
\title{undocumented internal method}
\usage{
check.votes.score(record, max.score, ...)
}
\arguments{
\item{record, max.score, ...}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
