% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{check.votes.stv}
\alias{check.votes.stv}
\title{undocumented internal method}
\usage{
check.votes.stv(record, equal.ranking = FALSE, ...)
}
\arguments{
\item{record, equal.ranking, ...}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
