% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{check.votes.tworound.runoff}
\alias{check.votes.tworound.runoff}
\title{undocumented internal method}
\usage{
check.votes.tworound.runoff(record, ...)
}
\arguments{
\item{record, ...}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
