% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRankings.R
\name{combineRankings}
\alias{combineRankings}
\title{the least upper bound on a pair of rankings}
\usage{
combineRankings(r1, r2)
}
\arguments{
\item{r1, r2}{numeric vectors}
}
\value{
the most complete (but possibly partial) ranking which is consistent
with both r1 and r2.  Uses 'ties.method="min"'
}
\description{
the least upper bound on a pair of rankings
}
\examples{
combineRankings( c(3,1,2), c(2,1,3) )
}
