% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{condorcet}
\alias{condorcet}
\title{Count votes using the Condorcet voting method.}
\usage{
condorcet(
  votes,
  runoff = FALSE,
  nseats = 1,
  safety = 1,
  fsep = "\\t",
  quiet = FALSE,
  ...
)
}
\arguments{
\item{votes}{[Matrix] or [data frame] containing the votes. Rows
correspond to the votes, columns correspond to the candidates. If `votes`
is a character string, it is interpreted as a file name from which the
votes are to be read. See below for more details.}

\item{runoff}{Logical. If [TRUE] and no Condorcet winner exists,
the election goes into a run-off, see below for details.}

\item{nseats}{the number of seats to be filled in this election}

\item{safety}{Parameter for a clustering heuristic on a total ranking of
the candidates.  Conjecture: the default of `1.0` ensures a separation
of one s.d. between clusters, when `votes` are i.u.d. permutations on the
candidates.}

\item{fsep}{If `votes` is a file name, this argument gives the column
separator in the file.}

\item{quiet}{If [TRUE] no output is printed.}

\item{...}{Additional arguments passed to the underlying functions. 
For the [image] function, see arguments for [image.SafeVote.stv], especially
[xpref], [ypref], [all.pref] and [proportion].}
}
\value{
Object of class SafeVote.condorcet
}
\description{
The Condorcet method elects the candidate who wins a majority of the ranked
vote in every head to head election against each of the other candidates.
*I.e.*, the Condorcet winner is a candidate who beats all other candidates
in pairwise comparisons. Analogously, a Condorcet loser is a candidate
who loses against all other candidates. Neither Condorcet winner nor loser
might exist.
}
\details{
If the runoff argument is set to `TRUE` and no Condorcet winner exists,
two or more candidates with the most pairwise wins are selected and
the method is applied to such subset. If more than two candidates are in
such run-off, the selection is performed repeatedly, until either a winner
is selected or no more selection is possible.

The input data votes is structured the same way as for the [stv] method: 
Row `i` contains the preferences of voter `i` numbered 
`1; 2; : : : ; r; 0; 0; 0; 0`, in some order, while equal preferences
are allowed. The columns correspond to the candidates. The [dimnames] of
the columns are the names of the candidates; if these are not supplied
then the candidates are lettered `A, B, C, ...`. If the dataset 
contains missing values ([NA]), they are replaced by zeros.

Note that if equal preferences are used, they are automatically converted
into a format where, for each preference $i$ which does not have any
duplicate, there must be exactly $i - 1$ preferences $j$ with $0 < j < i$. 
It is the same ranking as one would obtain with 
`rank(x, ties.method = "min")`. If a conversion of a vote occurs, 
a warning is issued. That is done internally by calling the
[correct.ranking] function.
}
\examples{
{
data(food_election)
condorcet(food_election)
}
}
