% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{correct.ranking}
\alias{correct.ranking}
\title{undocumented internal method}
\usage{
correct.ranking(votes, partial = FALSE, quiet = FALSE)
}
\arguments{
\item{votes, partial, quiet}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
