% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{.print.summary.SafeVote}
\alias{.print.summary.SafeVote}
\title{.print method for summary object}
\usage{
.print.summary.SafeVote(x, ...)
}
\arguments{
\item{x, ...}{undocumented}
}
\value{
undocumented
}
\description{
.print method for summary object
}
