% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{.summary.SafeVote}
\alias{.summary.SafeVote}
\title{summarises vote-totals for subsequent printing}
\usage{
.summary.SafeVote(object, larger.wins = TRUE, reorder = TRUE)
}
\arguments{
\item{object}{vector of total votes per candidate}

\item{larger.wins}{TRUE if candidates are "voted in" rather than voted-out}

\item{reorder}{TRUE if output data.frame columns should be in rank-order}
}
\value{
a data.frame with three columns and nc+1 rows, where
nc is the number of candidates.  The first column contains candidate
names and a final entry named "Sum".  The second column contains
vote totals. The third column is a vector of chars which indicate
whether the candidate has been elected.  The data.frame has four named
attributes carrying election parameters.

TODO: refactor into a modern dialect of R, perhaps by defining a
constructor for an election_info S3 object with a summary method and a
print method
}
\description{
summarises vote-totals for subsequent printing
}
