% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{electedAsRank}
\alias{electedAsRank}
\title{convert the names of elected candidates into a 0-1 ranking}
\usage{
electedAsRank(elected, candidates)
}
\arguments{
\item{elected}{the names of the elected candidates}

\item{candidates}{the names of all candidates}
}
\value{
a named 0-1 vector of length(candidates)
}
\description{
convert the names of elected candidates into a 0-1 ranking
}
