% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{election.info}
\alias{election.info}
\title{prints the basic results of an election}
\usage{
election.info(x)
}
\arguments{
\item{x}{basic election results, as named attributes of an R structure or
object}
}
\value{
data.frame : an invisible copy of the printed results

TODO: refactor into a modern dialect of R, e.g. defining a constructor for an
election_info S3 object with a print method
}
\description{
prints the basic results of an election
}
