% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{extractMargins}
\alias{extractMargins}
\title{extract margins from the results of a ballot count}
\usage{
extractMargins(marginNames, crRanks, cr)
}
\arguments{
\item{marginNames}{list of colnames of the margins in our SafeRank result}

\item{crRanks}{ranks of candidates, not necessarily total}

\item{cr}{structure returned by a ballot-counting method

Margins are adjusted for tied candidates, such that candidates within a tie
group have margins indicative of their relative strengths. Extremely small
margins are indicative of floating-point roundoff errors.}
}
\value{
named list of margins
}
\description{
extract margins from the results of a ballot count
}
