% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{extractRank}
\alias{extractRank}
\title{Extract a ranking vector by name from the results of a ballot count}
\usage{
extractRank(rankMethod, countMethod, cr)
}
\arguments{
\item{rankMethod}{e.g. "elected"}

\item{countMethod}{e.g. "stv"}

\item{cr}{structure returned by a ballot-counting method}
}
\value{
a numeric ranking vector, in order of colnames(cr$data)
}
\description{
Extract a ranking vector by name from the results of a ballot count
}
