% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{forwards.tiebreak}
\alias{forwards.tiebreak}
\title{Undocumented internal method}
\usage{
forwards.tiebreak(prefs, icans, elim = TRUE)
}
\arguments{
\item{prefs}{undocumented}

\item{icans}{undocumented}

\item{elim}{undocumented}
}
\description{
Undocumented internal method
}
