% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisborne2022}
\alias{gisborne2022}
\title{Results of Gisborne Local Elections 2022}
\format{
Results of \href{https://www.gdc.govt.nz/council/mayor-and-councillors/2022-local-elections/election-results}{three STV elections in Gisborne, New Zealand, in 2022}.
The results for each count of the ballots (progress, preliminary, and
final) include quotas, keep values, non-transferable votes(NTV), and the
number of the round in which each candidate was elected or eliminated.

A data frame with 1076 observations and 13 variables.  These are the
vote totals and rank-ordering of candidates in the progress and final
results of \href{https://www.aucklandcouncil.govt.nz/elections/elections-2022-results/Pages/default.aspx}{STV elections in Auckland, New Zealand, in 2022}.
"Progress results include votes returned up until Friday 7 October. Special
votes and votes hand delivered by 12 noon on Saturday 8 October are not
included... Final count includes all votes, including special votes that
are returned by 12 noon on Saturday 8 October."  We have calculated
nBallots from the votes as reported by the Electoral Officer, see
data-raw/auckland2022.R for details.  Curiously, the reported votes are all
integral, suggesting that some method other than Meek's was used to count
the ballots, or that some postprocessing of the usual reporting of Meek's
results was performed to avoid the (apparent) double-counting of votes
which occurs whenever an excluded candidate's votes are both reported
against this candidate \emph{and} also distributed among the candidates still in
play.
}
\usage{
data(gisborne2022)

data(auckland2022)
}
\description{
Results of Gisborne Local Elections 2022

Results of Auckland Local Elections 2022
}
\keyword{datasets}
