% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hamilton2022}
\alias{hamilton2022}
\title{Results of Hamilton Local Elections 2022}
\format{
\href{https://www.yourcityelections.co.nz/assets/Uploads/Results-page/Hamilton-City-Council-Progress-results-for-the-2022-Election.pdf}{Progress}
and \href{https://www.yourcityelections.co.nz/results/}{final} results for the
2022 local body elections in Hamilton New Zealand.  These results include
quotas, keep values, non-transferable votes(NTV), and the number of the
round in which each candidate was either elected or eliminated.
}
\usage{
data(hamilton2022)
}
\description{
Results of Hamilton Local Elections 2022
}
\keyword{datasets}
