% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ims_score}
\alias{ims_score}
\title{IMS Score}
\format{
A data frame with 620 observations and 10 candidates (names were
made up). Each record corresponds to one ballot, with higher values
indicating the more-preferred candidates.
}
\usage{
ims_score
}
\description{
Modified version of ims_election, for use in score voting.
}
\keyword{datasets}
