% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ims_stv}
\alias{ims_stv}
\title{IMS STV}
\format{
A data frame with 620 observations and 10 candidates (names were
made up). Each record corresponds to one ballot. The IMS Council voting
is done using the STV method, and thus the ims_election dataset contains
ballots with candidates being ranked between 1 and 10 with zeros allowed.
}
\usage{
ims_stv
}
\description{
Copy of ims_election, included for backwards compatibility.
}
\keyword{datasets}
