% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{invalid.votes}
\alias{invalid.votes}
\title{Extracts the invalid.votes member (if any) from the result of a count}
\usage{
invalid.votes(x)
}
\arguments{
\item{x}{value returned by stv, condorcet, approval, plurality, or score}
}
\value{
matrix with one column per candidate and one row per invalid
ballot
}
\description{
This method was added Jan 2022 -- it was named in a warning message but had
apparently either never been implemented, or had been "lost" through
versioning.
}
