% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{is.valid.vote}
\alias{is.valid.vote}
\title{undocumented internal method}
\usage{
is.valid.vote(x, method, ...)
}
\arguments{
\item{x, method, ...}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
