% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{ordered.tiebreak}
\alias{ordered.tiebreak}
\title{Undocumented internal method}
\usage{
ordered.tiebreak(vmat, seed = NULL)
}
\arguments{
\item{vmat}{undocumented}

\item{seed}{undocumented}
}
\description{
Undocumented internal method
}
