% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{plot.SafeVote.stv}
\alias{plot.SafeVote.stv}
\title{plot() method for the result of an stv() ballot-count}
\usage{
\method{plot}{SafeVote.stv}(x, xlab = "Count", ylab = "Preferences", point.size = 2, ...)
}
\arguments{
\item{x, xlab, ylab, point.size, ...}{undocumented}
}
\value{
undocumented
}
\description{
plot() method for the result of an stv() ballot-count
}
