% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{prepare.votes}
\alias{prepare.votes}
\title{undocumented internal method}
\usage{
prepare.votes(data, fsep = "\\n")
}
\arguments{
\item{data, fsep}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
