% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{print.summary.SafeRankExpt}
\alias{print.summary.SafeRankExpt}
\title{Print method for summary.SafeRankExpt}
\usage{
\method{print}{summary.SafeRankExpt}(x, ...)
}
\arguments{
\item{x}{experimental results}

\item{...}{args for generic print()}
}
\value{
invisible(x), with side-effects to console
}
\description{
Print method for summary.SafeRankExpt
}
