% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{print.summary.SafeVote.condorcet}
\alias{print.summary.SafeVote.condorcet}
\title{print method for summary.SafeVote.condorcet}
\usage{
\method{print}{summary.SafeVote.condorcet}(x, ...)
}
\arguments{
\item{x}{object of type summary.SafeVote.condorcet}

\item{...}{parameters passed to generic \link{print}}
}
\value{
textual description of 'x'
}
\description{
print method for summary.SafeVote.condorcet
}
