% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{print.summary.SafeVote.score}
\alias{print.summary.SafeVote.score}
\title{print method for summary.score object}
\usage{
\method{print}{summary.SafeVote.score}(x, ...)
}
\arguments{
\item{x, ...}{undocumented}
}
\value{
undocumented
}
\description{
print method for summary.score object
}
