% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{print.summary.SafeVote.stv}
\alias{print.summary.SafeVote.stv}
\title{print() method for a summary() of a SafeVote result}
\usage{
\method{print}{summary.SafeVote.stv}(x, ...)
}
\arguments{
\item{x}{election results}

\item{...}{args to be passed to kable()}
}
\value{
no return value, called for side-effect of printing to console
}
\description{
print() method for a summary() of a SafeVote result
}
