% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{rbind.SafeRankExpt}
\alias{rbind.SafeRankExpt}
\title{add a row to a SafeRankExpt object}
\usage{
\method{rbind}{SafeRankExpt}(object, row)
}
\arguments{
\item{object}{prior results of experimentation}

\item{row}{new observations}
}
\value{
updated SafeRankExpt object
}
\description{
add a row to a SafeRankExpt object
}
