% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{rbind_SafeRankExpt}
\alias{rbind_SafeRankExpt}
\title{add a row to a SafeRankExpt object, using dplyr::bind_rows()}
\usage{
rbind_SafeRankExpt(object, row)
}
\arguments{
\item{object}{prior results of experimentation}

\item{row}{new observations}
}
\value{
SafeRankExpt object with an additional row
}
\description{
add a row to a SafeRankExpt object, using dplyr::bind_rows()
}
