% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{remove.candidate}
\alias{remove.candidate}
\title{undocumented internal method}
\usage{
remove.candidate(votes, can, quiet = TRUE)
}
\arguments{
\item{votes, can, quiet}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
