% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{score}
\alias{score}
\title{Count votes using the score (or range) method.}
\usage{
score(
  votes,
  nseats = 1,
  max.score = NULL,
  larger.wins = TRUE,
  fsep = "\\t",
  quiet = FALSE,
  ...
)
}
\arguments{
\item{votes, nseats, max.score, larger.wins, fsep, quiet, ...}{undocumented}
}
\value{
undocumented
}
\description{
See https://arxiv.org/abs/2102.05801
}
