% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{solveTiebreak}
\alias{solveTiebreak}
\title{Undocumented internal method, renamed from 'solve.tiebreak' to avoid
confusion with generic solve()}
\usage{
solveTiebreak(method, prefs, icans, ordered.ranking = NULL, elim = TRUE)
}
\arguments{
\item{method}{undocumented}

\item{prefs}{undocumented}

\item{icans}{undocumented}

\item{ordered.ranking}{undocumented}

\item{elim}{undocumented}
}
\value{
undocumented
}
\description{
Undocumented internal method, renamed from 'solve.tiebreak' to avoid
confusion with generic solve()
}
