% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{summary.SafeRankExpt}
\alias{summary.SafeRankExpt}
\title{summary method for SafeRankExpt}
\usage{
\method{summary}{SafeRankExpt}(x, ...)
}
\arguments{
\item{x}{experimental results to be summarised}

\item{...}{args for generic summary()}
}
\value{
summary.SafeRankExpt object
}
\description{
summary method for SafeRankExpt
}
