% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{summary.SafeVote.approval}
\alias{summary.SafeVote.approval}
\title{summary method for approval results}
\usage{
\method{summary}{SafeVote.approval}(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
undocumented
}
\description{
summary method for approval results
}
