% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{summary.SafeVote.condorcet}
\alias{summary.SafeVote.condorcet}
\title{Summary method for condorcet() results}
\usage{
\method{summary}{SafeVote.condorcet}(object, ...)
}
\arguments{
\item{object}{of type SafeVote.condorcet}

\item{...}{undocumented, currently unused}
}
\value{
\link{data.frame} object
}
\description{
Summary method for condorcet() results
}
