% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{summary.SafeVote.plurality}
\alias{summary.SafeVote.plurality}
\title{summary method for plurality object}
\usage{
\method{summary}{SafeVote.plurality}(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
descriptive dataframe
}
\description{
summary method for plurality object
}
