% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{summary.SafeVote.score}
\alias{summary.SafeVote.score}
\title{summary method for score object}
\usage{
\method{summary}{SafeVote.score}(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
undocumented
}
\description{
summary method for score object
}
