% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{summary.vote.stv}
\alias{summary.vote.stv}
\title{summary() method for a backwards-compatible result}
\usage{
\method{summary}{vote.stv}(object, ..., digits = 3)
}
\arguments{
\item{object}{undocumented}

\item{...}{undocumented}

\item{digits}{undocumented}
}
\description{
summary() method for a backwards-compatible result
}
