% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{sumOfVotes}
\alias{sumOfVotes}
\title{undocumented internal method,}
\usage{
sumOfVotes(votes)
}
\arguments{
\item{votes}{undocumented}
}
\value{
undocumented
}
\description{
Renamed from 'sum.votes' to avoid confusion with the generic sum()
}
