% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{translate.ties}
\alias{translate.ties}
\title{Undocumented internal method from original code}
\usage{
translate.ties(ties, method)
}
\arguments{
\item{ties}{undocumented}

\item{method}{'f' for forward, 'b' for backward}
}
\value{
undocumented
}
\description{
Undocumented internal method from original code
}
