% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk_labour_2010}
\alias{uk_labour_2010}
\title{UK Labour Party Leader 2010}
\format{
A data frame with 266 observations and 5 candidates.
}
\usage{
data(uk_labour_2010)
}
\description{
These are the ballots cast by Labour MPs and MEPs in an election
of their party's leader in 2010, as published by the Manchester Guardian.
The names of the electors have been suppressed in this file, but are
available at \href{https://rangevoting.org/LabourUK2010.html}{Rangevoting.org}
along with extensive commentary on the election.
}
\keyword{datasets}
