% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{view}
\alias{view}
\title{generic view() for classes defined in this package}
\usage{
view(object, ...)
}
\arguments{
\item{object}{election object to be viewed}

\item{...}{additional parameters, passed to formattable::formattable()}
}
\value{
html-formatted object, with side-effect in RStudio's Viewer pane
}
\description{
generic view() for classes defined in this package
}
