% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{View.SafeVote.approval}
\alias{View.SafeVote.approval}
\title{View method for approval object}
\usage{
View.SafeVote.approval(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
undocumented
}
\description{
View method for approval object
}
