% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{View.SafeVote.condorcet}
\alias{View.SafeVote.condorcet}
\title{View method for SafeVote.condorcet}
\usage{
View.SafeVote.condorcet(object, ...)
}
\arguments{
\item{object}{of type SafeVote.condorcet}

\item{...}{see \link{View.SafeVote.approval}}
}
\value{
View object
}
\description{
View method for SafeVote.condorcet
}
