% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{View.SafeVote.plurality}
\alias{View.SafeVote.plurality}
\title{View method for plurality object}
\usage{
View.SafeVote.plurality(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
View object
}
\description{
View method for plurality object
}
