% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{view.SafeVote.score}
\alias{view.SafeVote.score}
\title{view method for score object}
\usage{
\method{view.SafeVote}{score}(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
undocumented
}
\description{
view method for score object
}
