% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{View.SafeVote.stv}
\alias{View.SafeVote.stv}
\title{View method for the result of an stv() ballot-count}
\usage{
View.SafeVote.stv(object, ...)
}
\arguments{
\item{object}{object to be viewed}

\item{...}{additional parameters, passed to formattable::formattable()}
}
\value{
html-formatted object
}
\description{
View method for the result of an stv() ballot-count
}
