% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wellington2022analysis}
\alias{wellington2022analysis}
\title{Analysis of results from the Wellington 2022 election.}
\format{
a dataframe with a column labelled "closeness" -- which is easily
calculated and might be a valid metric for the safety of the results of an
STV election.
}
\usage{
wellington2022analysis
}
\description{
Analysis of results from the Wellington 2022 election.
}
\keyword{datasets}
