% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{as.SafeRankExpt}
\alias{as.SafeRankExpt}
\title{as.SafeRankExpt()}
\usage{
as.SafeRankExpt(df)
}
\arguments{
\item{df}{data.frame object}
}
\value{
a SafeRankExpt object, or stop() if df fails some sanity checks
}
\description{
as.SafeRankExpt()
}
