% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{completeRankingTable}
\alias{completeRankingTable}
\title{internal method to analyse the partial results of an stv() ballot count,
to discover a complete ranking of all candidates.  The ranking may depend
on the value of nseats, because this affects how votes are transferred.}
\usage{
completeRankingTable(object, quiet, verbose)
}
\arguments{
\item{object}{partial results}

\item{quiet}{TRUE to suppress console output}

\item{verbose}{TRUE to produce diagnostic output}
}
\value{
data.frame with columns Rank, Margin, Candidate, Elected, SafeRank
}
\description{
internal method to analyse the partial results of an stv() ballot count,
to discover a complete ranking of all candidates.  The ranking may depend
on the value of nseats, because this affects how votes are transferred.
}
