% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{.summary.SafeVote}
\alias{.summary.SafeVote}
\title{.summary method for objects in SafeRank class}
\usage{
.summary.SafeVote(object, larger.wins = TRUE, reorder = TRUE)
}
\arguments{
\item{object, larger.wins, reorder}{undocumented}
}
\value{
undocumented
}
\description{
.summary method for objects in SafeRank class
}
