% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{election.info}
\alias{election.info}
\title{undocumented internal method}
\usage{
election.info(x)
}
\arguments{
\item{x}{undocumented}
}
\value{
undocumented
}
\description{
undocumented internal method
}
