% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{extractMargins}
\alias{extractMargins}
\title{extract margins from the results of a ballot count}
\usage{
extractMargins(marginNames, rankMethod, cr)
}
\arguments{
\item{marginNames}{list of colnames of the margins in our SafeRank result}

\item{rankMethod}{if "safeRank", margins are adjusted appropriately}

\item{cr}{structure returned by a ballot-counting method}
}
\value{
named list of margins
}
\description{
extract margins from the results of a ballot count
}
