% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_election}
\alias{food_election}
\title{Food Election}
\format{
A data frame with 20 observations and 5 candidates (Oranges,
Pears, Chocolate, Strawberries, Sweets). Each record corresponds to
one ballot with ranking for each of the candidates.
}
\usage{
data(food_election)
}
\description{
Sample data for testing SafeVote
}
\keyword{datasets}
