% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{image.SafeVote.condorcet}
\alias{image.SafeVote.condorcet}
\title{The image function visualizes the joint distribution of two preferences
(if \code{all.pref=FALSE}) given \code{xpref} and \code{ypref}, as well as the marginal
distribution of all preferences (if \code{all.pref=TRUE}). The joint distribution
can be shown as proportions (if \code{proportion=TRUE}) or raw vote counts
(if \code{proportion=FALSE}).}
\usage{
\method{image}{SafeVote.condorcet}(x, ...)
}
\arguments{
\item{x}{object of type SafeVote.condorcet}

\item{...}{See arguments for \link{image.SafeVote.stv}, especially \code{xpref},
\code{ypref}, \code{all.pref} and \code{proportion}.}
}
\value{
image object, with side-effect in RStudio Plots pane
}
\description{
The image function visualizes the joint distribution of two preferences
(if \code{all.pref=FALSE}) given \code{xpref} and \code{ypref}, as well as the marginal
distribution of all preferences (if \code{all.pref=TRUE}). The joint distribution
can be shown as proportions (if \code{proportion=TRUE}) or raw vote counts
(if \code{proportion=FALSE}).
}
