% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{image.SafeVote.stv}
\alias{image.SafeVote.stv}
\title{image() method for the result of an stv() ballot-count}
\usage{
\method{image}{SafeVote.stv}(x, xpref = 2, ypref = 1, all.pref = FALSE, proportion = TRUE, ...)
}
\arguments{
\item{x, xpref, ypref, all.pref, proportion, ...}{undocumented (legacy code)}
}
\value{
image object, with side-effect in RStudio Plots pane
}
\description{
image() method for the result of an stv() ballot-count
}
