% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ims_approval}
\alias{ims_approval}
\title{IMS Approval}
\format{
A data frame with 620 observations and 10 candidates (names were 
made up). Each record corresponds to one ballot, with 0 indicating
disapproval of a candidate and 1 indicating approval.
}
\usage{
data(ims_approval)
}
\description{
Modified version of ims_election, for use in approval voting.
}
\keyword{datasets}
