% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ims_plurality}
\alias{ims_plurality}
\title{IMS Plurality}
\format{
A data frame with 620 observations and 10 candidates (names were 
made up). Each record corresponds to one ballot, with 1 against
the voter's most-preferred candidate and 0 against all other candidates.
}
\usage{
data(ims_plurality)
}
\description{
Modified version of ims_election, for use in plurality voting.
}
\keyword{datasets}
