% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{is.SafeRankExpt}
\alias{is.SafeRankExpt}
\title{is.SafeRankExpt()}
\usage{
is.SafeRankExpt(x)
}
\arguments{
\item{x}{object of unknown class}
}
\value{
TRUE if x is a SafeRankExpt object
}
\description{
Throws an error if the \code{.internal.selfref} sentinel indicates that
the underlying \code{data.table} has been corrupted by copying.
}
