% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{loserMargin}
\alias{loserMargin}
\title{Find a loser and their margin of victory}
\usage{
loserMargin(votes)
}
\arguments{
\item{votes}{cleaned ballots}
}
\value{
length-2 vector: the index of a losing candidate, and their margin
        of loss (0 if a tie, NA if no winners)
}
\description{
Find a loser and their margin of victory
}
