% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{new_SafeRankExpt}
\alias{new_SafeRankExpt}
\title{Constructor for the results of a SafeRank experiment}
\usage{
new_SafeRankExpt(
  nvotes = vector(mode = integer),
  Ranks = matrix(),
  Margins = matrix(),
  countMethod = str(),
  rankMethod = str(),
  OtherFactors = vector()
)
}
\arguments{
\item{nvotes}{primary factor: number of ballots counted}

\item{Ranks}{observed ranks}

\item{Margins}{observed margins}

\item{countMethod}{secondary factor: counting method e.g. "stv"}

\item{rankMethod}{secondary factor: ranking method e.g. "elected"}

\item{OtherFactors}{other secondary factors e.g. dataset name}
}
\value{
object of class SafeRankExpt
}
\description{
Constructor for the results of a SafeRank experiment
}
