% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{new_SafeRankExpt}
\alias{new_SafeRankExpt}
\title{Constructor for the results of a SafeRank experiment}
\usage{
new_SafeRankExpt(
  nBallots = integer(0),
  ranks = matrix(),
  margins = matrix(),
  countMethod = NULL,
  rankMethod = NULL,
  datasetName = NULL,
  experimentalMethod = NULL,
  otherFactors = list()
)
}
\arguments{
\item{nBallots}{primary factor: number of ballots counted in this
experimental unit}

\item{ranks}{a matrix with colnames(ballots), one row per experimental unit}

\item{margins}{a matrix with colnames(ballots), one row per unit}

\item{countMethod}{secondary factor: counting method e.g. "stv"}

\item{rankMethod}{secondary factor: ranking method e.g. "elected"}

\item{datasetName}{secondary factor: name of the dataset of ballots}

\item{experimentalMethod}{secondary factor: name of the method which
simulated these elections e.g. "testFraction"}

\item{OtherFactors}{other secondary factors, e.g. parameters to
experimentalMethod, a timestamp.}
}
\value{
object of class SafeRankExpt
}
\description{
Constructor for the results of a SafeRank experiment
}
