% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{plot.SafeRankExpt}
\alias{plot.SafeRankExpt}
\title{plot() method for the result of an experiment with varying numbers of ballots}
\usage{
\method{plot}{SafeRankExpt}(x, xlab = "Ballots", ylab = "Score", point.size = 2, ...)
}
\arguments{
\item{x}{object containing experimental results}

\item{xlab, ylab}{axis labels}

\item{point.size}{diameter of elected/eliminated points}

\item{...}{params for generic plot()}
}
\value{
graphics object, with side-effect in RStudio's Plots pane

  The score of a candidate is their ranking plus their scaled "winning
  margin".  The scaling factor is \eqn{\sqrt{n}/4}, where \eqn{n} is the
  number of ballots in this simulated election.

  Todo: amend scoring so that the margin is visually at the "top" of the
  tier. 
  
  Todo: extend to visualise results of testFraction(..., ainc=0)
}
\description{
plot() method for the result of an experiment with varying numbers of ballots
}
