% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{plot.SafeRankExpt}
\alias{plot.SafeRankExpt}
\title{plot() method for the result of an experiment with varying numbers of ballots}
\usage{
\method{plot}{SafeRankExpt}(
  x,
  facetWrap = FALSE,
  cMargin = 0.25,
  xlab = "Ballots",
  ylab = "Adjusted Rank",
  point.size = 2,
  ...
)
}
\arguments{
\item{x}{object containing experimental results}

\item{facetWrap}{TRUE provides per-candidate scatterplots}

\item{cMargin}{adjustable parameter in scoring}

\item{xlab, ylab}{axis labels}

\item{point.size}{diameter of elected/eliminated points}

\item{...}{params for generic plot()}
}
\value{
graphics object, with side-effect in RStudio Plots pane
}
\description{
The score of a candidate is their ranking plus their scaled "winning margin".
The scaled margin is \eqn{e^{-cx/\sqrt{n}}}, where \eqn{x} is the
unscaled margin (i.e. the number of votes by which this candidate is ahead of
the next-weaker candidate), \eqn{n} is the number of ballots in this
simulated election, and \eqn{c} is the scaling parameter `cmargin`.
}
\details{
Todo: invert the y axis so that the higher-ranked candidates are the "cream"
rather than the "dregs" on the chart.

Todo: list candidates in order of final ranking

Todo: amend so that testFraction(..., ainc=0) is a 2-d plot.
}
