% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{plot.SafeRankExpt}
\alias{plot.SafeRankExpt}
\title{plot() method for the result of an experiment with varying numbers of ballots}
\usage{
\method{plot}{SafeRankExpt}(
  x,
  facetWrap = FALSE,
  cMargin = 1,
  xlab = "Ballots",
  ylab = "Adjusted Rank",
  pointSize = 1,
  ...
)
}
\arguments{
\item{x}{object containing experimental results}

\item{facetWrap}{TRUE provides per-candidate scatterplots}

\item{cMargin}{parameter in rank-adjustment formula}

\item{xlab, ylab}{axis labels}

\item{pointSize}{diameter of points}

\item{...}{params for generic plot()}
}
\value{
graphics object, with side-effect in RStudio Plots pane
}
\description{
The score of a candidate is their ranking plus their scaled "winning margin".
The scaled margin is \eqn{e^{-cx/\sqrt{n}}}, where \eqn{x} is the
unscaled margin (i.e. the number of votes by which this candidate is ahead of
the next-weaker candidate), \eqn{n} is the number of ballots in this
simulated election, and \eqn{c} is the scaling parameter \code{cmargin}.
}
\details{
The default value of cmargin = 1.0 preserves the slope of score-trends across
rank-boundaries.  Smaller values of cmargin make it easier to see the
statistical fluctuations in margins as a function of nBallots.  cmargin = 0
displays only the ranks.

Todo: accept a list of experimental results.  Multiple runs with the same
number of ballots should be shown with a box-and-whisker rather than a point.
}
