% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{plot.SafeVote.stv}
\alias{plot.SafeVote.stv}
\title{plot() method for the result of an stv() ballot-count}
\usage{
\method{plot}{SafeVote.stv}(x, xlab = "Count", ylab = "Preferences", point.size = 2, ...)
}
\arguments{
\item{x}{stv results}

\item{xlab, ylab}{axis labels}

\item{point.size}{diameter of elected/eliminated points}

\item{...}{params for generic plot()}
}
\value{
graphics object, with side-effect in RStudio's Plots pane
}
\description{
The \code{plot} function shows the evolution of the total score for each candidate
as well as the quota.
}
