% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{prepare.votes}
\alias{prepare.votes}
\title{Coerce input 'data' into a matrix 'votes'}
\usage{
prepare.votes(data, fsep = "\\n")
}
\arguments{
\item{data}{possibly a .csv file, possibly an R object}

\item{fsep}{separation character for .csv e.g. tab or comma}
}
\value{
a matrix with one row per ballot, one column per candidate, with
  named rows and columns
}
\description{
Coerce input 'data' into a matrix 'votes'
}
