% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_votes.R
\name{remove.candidate}
\alias{remove.candidate}
\title{Remove a candidate, amending ballot papers as required}
\usage{
remove.candidate(votes, can, quiet = TRUE)
}
\arguments{
\item{votes}{ballot box}

\item{can}{candidate to be removed}

\item{quiet}{suppress diagnostics}
}
\value{
amended ballot box
}
\description{
Remove a candidate, amending ballot papers as required
}
