% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeVote-package.R
\docType{package}
\name{SafeVote-package}
\alias{SafeVote}
\alias{SafeVote-package}
\title{SafeVote: Election Vote Counting with Safety Features}
\description{
Experimental fork of vote_2.3-2, with SafeRank and safety tests.
}
\keyword{internal}
