% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{stv}
\alias{stv}
\title{Count preferential ballots using an STV method}
\usage{
stv(
  votes,
  nseats = NULL,
  eps = 0.001,
  equal.ranking = FALSE,
  fsep = "\\t",
  ties = c("f", "b"),
  constant.quota = FALSE,
  quota.hare = FALSE,
  group.nseats = NULL,
  group.members = NULL,
  complete.ranking = FALSE,
  invalid.partial = FALSE,
  verbose = FALSE,
  seed = NULL,
  quiet = FALSE,
  digits = 3,
  backwards.compatible = FALSE,
  safety = 1,
  ...
)
}
\arguments{
\item{votes}{an array with one column per candidate and one row per ballot,
as described in the \href{condorcet.html#details}{Details for condorcet()}}

\item{nseats}{the number of seats to be filled in this election}

\item{eps}{fuzz-factor when comparing fractional votes}

\item{equal.ranking}{= FALSE,}

\item{fsep}{column-separator for output}

\item{ties}{vector of tie-breaking methods: \code{'f'} for forward, \code{'b'} for
backward}

\item{constant.quota}{\code{TRUE} if quota is held constant.  Over-rides
quota.hare. Default is \code{FALSE}.}

\item{quota.hare}{\code{TRUE} if Hare quota, \code{FALSE} if Droop quota (default)}

\item{group.nseats}{number of seats reserved to members of a group, as in the
Church of England's STV methodology.}

\item{group.members}{vector of members of a group with reserved seats}

\item{complete.ranking}{\code{TRUE} (default) if a complete ranking of all
candidates will be produced.  This parameter affects the value assigned to
\code{nseats} when \code{\link[=stv]{stv()}} is called with \code{nseats=NULL}, but has no effect on
elections in which \code{nseats} is explicitly specified.}

\item{invalid.partial}{\code{TRUE} if ballots which do not specify a complete
ranking of candidates are informal (aka "invalid") \emph{i.e.} ignored
(with a warning).  Default is \code{FALSE}.}

\item{verbose}{\code{TRUE} for diagnostic output}

\item{seed}{integer seed for tie-breaking.  Warning: if non-\code{NULL}, the PRNG
for R is reseeded prior to \emph{every} random tie-break among the
possibly-elected candidates.  We have preserved this functionality in this
branch to allow regression against the legacy codebase of
\href{https://cran.r-project.org/web/packages/vote/index.html}{vote_2.3.2}. In
\link{SafeVote} the default value for seed is \code{NULL} rather than the legacy
value of 1234, to mitigate the validity hazard of PRNG reseedings during a
stochastic experiment.}

\item{quiet}{\code{TRUE} to suppress console output}

\item{digits}{number of significant digits in the output table}

\item{backwards.compatible}{\code{TRUE} to regress against vote2_3.2 by
disabling $margins, $fuzz, $rankingTable, $safeRank}

\item{safety}{number of standard deviations on vote-counts, when producing a
safeRank by clustering near-ties in a complete ranking}

\item{...}{undocumented intent (preserved from legacy code)}
}
\value{
object of class \code{vote.stv}.  Note: the winning margins in this object
are valid for the elected candidates and their (total) ranking, but must be
adjusted within tiegroups to be valid for the candidates' (possibly
partial) safeRank.
}
\description{
The \code{votes} parameter is as described in the \href{condorcet.html#details}{Details for condorcet()} with the following additional semantics.
If a ballot fails to rank all candidates, it is informal (\emph{i.e.} "invalid")
unless \code{invalid.partial=FALSE}. If a ballot has equally-ranked candidates, it
is informal unless \code{equal.preferences=TRUE}. Informal ballots are removed
from the ballot box before counting, a warning is issued,
and the ballots
}
\examples{
data(food_election)
stv(food_election, safety = 0.0)
stv(food_election, nseats = 2)

}
