% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{stv}
\alias{stv}
\title{Count preferential ballots using an STV method}
\usage{
stv(
  votes,
  nseats = NULL,
  eps = 0.001,
  equal.ranking = FALSE,
  fsep = "\\t",
  ties = c("f", "b"),
  constant.quota = FALSE,
  quota.hare = FALSE,
  group.nseats = NULL,
  group.members = NULL,
  complete.ranking = FALSE,
  invalid.partial = FALSE,
  verbose = FALSE,
  seed = 1234,
  quiet = FALSE,
  digits = 3,
  safety = 1,
  ...
)
}
\arguments{
\item{votes}{an array with one column per candidate and one row per ballot}

\item{nseats}{the number of seats to be filled in this election}

\item{eps}{fuzz-factor when comparing fractional votes}

\item{equal.ranking}{TRUE if ballots are allowed to rank candidates equally}

\item{fsep}{column-separator for output}

\item{ties}{vector of tie-breaking methods: 'f' for forward, 'b' for backward}

\item{constant.quota}{TRUE if quota is held constant.  Over-rides quota.hare.}

\item{quota.hare}{TRUE if Hare quota, FALSE if Droop quota}

\item{group.nseats}{number of seats reserved to members of a group, as in the
Church of England's STV methodology.}

\item{group.members}{vector of members of a group with reserved seats}

\item{complete.ranking}{TRUE if the balloting is intended to produce a
ranking of all candidates.  This affects the value assigned to nseats when
stv() is called with nseats=NULL}

\item{invalid.partial}{TRUE if ballots are invalid (aka "informal") if they
do not specify a complete ranking of candidates}

\item{verbose}{TRUE for diagnostic output}

\item{seed}{integer seed for the (default) RNG in this instance of R}

\item{quiet}{TRUE to suppress console output}

\item{digits}{number of significant digits in the output table}

\item{safety}{number of standard deviations on vote-counts, when producing a
safeRank by clustering near-ties in a complete ranking}

\item{...}{undocumented intent}
}
\value{
object of class vote.stv
}
\description{
Count preferential ballots using an STV method
}
\examples{
data(food_election)
stv(food_election, complete.ranking=TRUE, safety=0.5)
stv(food_election, nseats = NULL, eps = 0.001,
  equal.ranking = FALSE, fsep = "\t", ties = c("f", "b"),
  constant.quota = FALSE, quota.hare = FALSE,
  group.nseats = NULL, group.members = NULL,
  complete.ranking = FALSE, invalid.partial = FALSE,
  verbose = FALSE, seed = 1234, quiet = FALSE, digits = 3)

}
