% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{summary.SafeVote.stv}
\alias{summary.SafeVote.stv}
\title{summary() method for a SafeVote result}
\usage{
\method{summary}{SafeVote.stv}(object, ..., digits = 3)
}
\arguments{
\item{object}{undocumented}

\item{...}{undocumented}

\item{digits}{undocumented}
}
\description{
summary() method for a SafeVote result
}
