% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{sumOfVotes}
\alias{sumOfVotes}
\title{internal method, computes column-sums}
\usage{
sumOfVotes(votes)
}
\arguments{
\item{votes}{ballots are rows, candidates are columns}
}
\value{
vector of votes for each candidate
}
\description{
Renamed from 'sum.votes' to avoid confusion with the generic sum()
}
