% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{testDeletions}
\alias{testDeletions}
\title{Assess the safety of a preliminary result for an election}
\usage{
testDeletions(
  votes = "food_election",
  countMethod = "stv",
  countArgs = list(),
  dstart = NULL,
  dinc = NULL,
  dlimit = NULL,
  drep = NULL,
  rankMethod = "safeRank",
  exptName = NULL,
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{votes}{A set of ballots}

\item{countMethod}{The name of a function which will count the ballots}

\item{countArgs}{List of args to be passed to countMethod (in addition to
votes)}

\item{dstart}{Number of ballots in the first ballot-count (selected at random
from votes, without replacement)}

\item{dinc}{Number of ballots to be deleted in subsequent steps}

\item{dlimit}{Maximum number of ballots to delete (in addition to dstart)}

\item{drep}{Maximum number of elections (required if dinc=0)}

\item{rankMethod}{Name of a ranking attribute in the output of countMethod}

\item{exptName}{stem-name of experimental units e.g. "E"}

\item{quiet}{TRUE to suppress all output}

\item{verbose}{TRUE to produce diagnostic output}
}
\value{
SafeRankExpt object, describing this experiment and its results
}
\description{
Ballots are deleted at random from the ballot-box, with election results
computed once per 'dinc' ballot-deletions.  The experiment terminates after a
specified number of ballots have been deleted, or a specified number of
ballot-counts have occurred.  Note: these ballot-counts are correlated.  Use
testFraction() to experiment with independently-drawn samples from the
ballot-box.
}
\examples{
data(food_election)
testDeletions(food_election)
testDeletions(food_election, countMethod="stv",
  countArgs=list(complete.ranking=TRUE))

}
