% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{testDeletions}
\alias{testDeletions}
\title{Test the sensitivity of a ranking to random deletions}
\usage{
testDeletions(
  votes = "food_election",
  countMethod = "condorcet",
  countArgs = NULL,
  dlimit = NULL,
  dinc = NULL,
  rankMethod = "SafeRank"
)
}
\arguments{
\item{votes}{A set of ballots}

\item{countMethod}{The name of a function which will count the
ballots}

\item{countArgs}{List of args to be passed to countMethod (in
addition to votes)}

\item{dlimit}{Maximum number of ballots to delete}

\item{dinc}{Number of ballots to be deleted in each step}

\item{rankMethod}{Name of a ranking attribute in the output of
countMethod}
}
\value{
A vector of the indices of ballots whose deletions affected
    the ranking. The vector is of zero length if the 'dlimit'
    termination condition had been reached without affecting the
    ranking.
}
\description{
Ballots are deleted until the ranking changes, or
until a specified number of ballots (`dlimit`) have been deleted.
}
\examples{
{
data(food_election)
testDeletions(food_election, countMethod="condorcet", dlimit=4)
testDeletions(food_election, countMethod="stv", countArgs=list(nseats=10))
}
}
