% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{testFraction}
\alias{testFraction}
\title{Test the fraction of ballots required to determine a result.}
\usage{
testFraction(
  votes,
  astart = NULL,
  ainc = NULL,
  arep = NULL,
  rankMethod = "SafeRank",
  countMethod = "stv",
  countArgs = NULL
)
}
\arguments{
\item{votes}{A set of ballots}

\item{astart}{Starting number of ballots (min 2)}

\item{ainc}{Number of ballots to be added in each step}

\item{arep}{Limit on the number of repetitions of the test.
Required to be non-null if ainc==0.}

\item{rankMethod}{Name of a ranking attribute in the output of
countMethod}

\item{countMethod}{The name of a function which will count the
ballots}

\item{countArgs}{List of args to be passed to countMethod (in
addition to votes)}
}
\value{
Number of ballots required to reach the same result as when
    all the ballots are counted.  The final ballot count is
    attached as an attribute "finalResults".
}
\description{
Starting from some number ('astart') of randomly-selected ballots,
additional randomly-selected ballots are added until the same
result is obtained as when all of the ballots are counted.
}
\examples{
data(food_election)
testFraction(food_election, countMethod="condorcet", 
              countArgs=list(safety=0.5))

}
