% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferank.R
\name{testFraction}
\alias{testFraction}
\title{Bootstrapping experiment, with fractional counts of a ballot box.}
\usage{
testFraction(
  votes,
  astart = NULL,
  ainc = NULL,
  arep = NULL,
  rankMethod = "safeRank",
  countMethod = "stv",
  countArgs = list(),
  exptName = NULL,
  quiet = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{votes}{A numeric matrix: one row per ballot, one column per candidate}

\item{astart}{Starting number of ballots (min 2)}

\item{ainc}{Number of ballots to be added in each step}

\item{arep}{Limit on the number of repetitions of the test. Required to be
non-null if ainc==0.}

\item{rankMethod}{Name of a ranking attribute in the output of countMethod,
e.g. "elected", "ranking", "safeRank".}

\item{countMethod}{The name of a function which will count the ballots, e.g.
"stv", "condorcet"}

\item{countArgs}{List of args to be passed to countMethod (in addition to
votes)}

\item{exptName}{stem-name of experimental units e.g. "E"}

\item{quiet}{TRUE to suppress all output}

\item{verbose}{TRUE to produce diagnostic output}
}
\value{
a matrix of experimental results, of dimension n by 2m+1, where n is
  the number of elections and m is the number of candidates.  The first
  column is named "nBallots".  Other columns indicate the ranking and margin
  of each candidate in each election.
}
\description{
Starting from some number ('astart') of randomly-selected ballots, an
increasingly-large collection of randomly-selected ballots are counted. The
ballots are chosen independently without replacement for each experimental
unit; if you want to count decreasingly-sized portions of a single sample of
ballots, use testDeletions().  The rankings and margins produced by each
simulated election are returned as a matrix with one row per election.
}
\examples{
data(food_election)
testFraction(food_election, countMethod="condorcet",
             countArgs=list(safety=0.5,complete.ranking=TRUE))
testFraction(dublin_west, astart=20, ainc=20, arep=2, countMethod="stv",
             rankMethod="elected", quiet=FALSE)
}
