% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{view.SafeVote.approval}
\alias{view.SafeVote.approval}
\title{view method for approval object}
\usage{
view.SafeVote.approval(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
undocumented
}
\description{
view method for approval object
}
