% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet.R
\name{view.SafeVote.condorcet}
\alias{view.SafeVote.condorcet}
\title{view method for SafeVote.condorcet}
\usage{
\method{view}{SafeVote.condorcet}(object, ...)
}
\arguments{
\item{object}{of type SafeVote.condorcet}

\item{...}{see \link{view.SafeVote.approval}}
}
\value{
view object
}
\description{
view method for SafeVote.condorcet
}
