% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_by_sum.R
\name{view.SafeVote.plurality}
\alias{view.SafeVote.plurality}
\title{view method for plurality object}
\usage{
view.SafeVote.plurality(object, ...)
}
\arguments{
\item{object, ...}{undocumented}
}
\value{
undocumented
}
\description{
view method for plurality object
}
