% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{view.SafeVote.stv}
\alias{view.SafeVote.stv}
\title{view method for the result of an stv() ballot-count}
\usage{
\method{view}{SafeVote.stv}(object, ...)
}
\arguments{
\item{object}{object to be viewed}

\item{...}{additional parameters, passed to formattable::formattable()}
}
\value{
html-formatted object
}
\description{
view method for the result of an stv() ballot-count
}
