% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stv.R
\name{winnerMargin}
\alias{winnerMargin}
\title{Find a winner and their margin of victory}
\usage{
winnerMargin(votes)
}
\arguments{
\item{votes}{cleaned ballots}
}
\value{
length-2 vector: the index of a winning candidate, and their margin
of victory (0 if a tie, NA if no losers)
}
\description{
Find a winner and their margin of victory
}
